+++ src/api/auth.ts
@@ -1,7 +1,7 @@
 import { LoginFormData } from '@/interfaces/LoginFormData';
 import { SignUpFormData } from '@/interfaces/SignUpFormData';
 
-import axiosInstance from '@/api/apiInstance';
+import axiosInstance from '@/api/axiosInstance';
 
 const login = async (data: LoginFormData) => {
   try {
